Frame Minimum Size
If the user resizes the window down too far, it creates a “slinky”
effect with all of the components inside the application, making them
shrink and distort. This can be an unattractive UI feature. I added a
method that sets the minimum size of the window.

I chose the minimum size based on what size was the smallest possible
dimension while not distorting the most complex page in the application
(the Edit Appointment page). That way all pages look good.

I have attached two videos to demonstrate the app without my changes
and with them. Any feedback is welcome, as always.

Thanks,
Patrick